GUI: darken background somewhat
Reduce the lightness of the background color for GUI mode, with
corresponding adjustment to the background color of the `CursorLine`.

The motivation for this change is that I found the background color of
the `CursorLine` too light, such that syntax-highlighted text often
failed to contrast with it sufficiently to be easily legible to me,
and reducing the lightness of the overall background color in addition
to that of the `CursorLine` background color was necessary for the
`CursorLine` to remain distinct.