Remove KWEasy in favor of EasyMotion
I have concluded that I prefer the plugin EasyMotion over the similar
plugin KWEasy. Both provide mappings for searching for a character or
string in the open buffer and then selecting a match to jump to by
pressing a key, but whereas EasyMotion simply highlights the matches
while greying out the rest of the searched text, KWEasy blanks out the
rest of the text and shows only the choices to jump to (and does so
buggily, without seeming to properly account for the effects of the
'showbreak' option), which I find disorienting. Additionally,
EasyMotion is more featureful, which I fail to resent.

Therefore, remove KWEasy from the set of my enabled plugins.