RecolorLineNrsForMode: Take more care re cterm/gui
In the function `RecolorLineNrsForMode`, before this change, when
changing the syntax highlighting settings for line-numbers as
considered appropriate for Normal mode, the line-numbers' `ctermfg`
and `guifg` syntax highlighting attributes would both be set to the
same value, regardless of whether said value was applicable to cterm
or GUI mode.

Vim seemed to be okay with this, but now Vim 8 seems to have become
more strict, such that this results in errors being raised.

This change seemingly fixes this problem by only setting the `ctermfg`
xor `guifg` attribute in the afore-described functionality, selecting
the attribute appropriate to the Vim instance's current display mode.