Add fn StrToNrWrapping
The standard VimScript function `str2nr` has changed from wrapping to
saturating in its calculation of its result when the number encoded in
the input is larger than the output (a 64-bit signed integer) can
hold.

This patch adds a function `StrToNrWrapping` to replicate the old
behavior of `str2nr`.